+++ com.unity.render-pipelines.high-definition/Editor/Material/Lit/BaseLitGUI.cs
@@ -189,6 +189,39 @@ namespace UnityEditor.Rendering.HighDefinition
         static public void SetupBaseLitMaterialPass(Material material)
         {
             material.SetupBaseUnlitPass();
+
+            bool emissiveIsActive = false;
+            bool emissionPassCanBeDisabled = false;
+            if (material.HasProperty(kUseEmissiveIntensity))
+            {
+                emissionPassCanBeDisabled = true;
+                var useIntensity = material.GetInt(kUseEmissiveIntensity);
+                if (useIntensity == 0)
+                {
+                    if (material.HasProperty(kEmissiveColor))
+                    {
+                        var emissionColor = material.GetColor(kEmissiveColor);
+                        if (emissionColor.r > 0.0 || emissionColor.g > 0.0 || emissionColor.b > 0.0)
+                        {
+                            emissiveIsActive = true;
+                        }
+                    }
+                }
+                else
+                {
+                    if (material.HasProperty(kEmissiveIntensity))
+                    {
+                        var intensityValue = material.GetFloat(kEmissiveIntensity);
+                        if (intensityValue > 0.0)
+                        {
+                            emissiveIsActive = true;
+                        }
+                    }
+                }
+            }
+
+            if (emissionPassCanBeDisabled)
+                material.SetShaderPassEnabled(HDShaderPassNames.s_ForwardEmissiveStr, emissiveIsActive);
         }
     }
 } // namespace UnityEditor